# code/financial_planner_api_scenario/openapi/financial-services.yaml
openapi: 3.0.3
info:
  title: Financial Services Mini App API
  version: 0.2.0
  description: >
    OpenAPI contract for the financial services v2 mini-app. This API models clients,
    households, investments, savings, superannuation, insurance, communications, campaigns,
    tasks, anniversaries, tax checkpoints, and admin/dashboard operations.
servers:
  - url: http://localhost:8082
...
paths:
...
  /v1/clients:
    get:
      operationId: listClients
      summary: List clients
      tags: [clients]
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
    post:
      operationId: createClient
      summary: Create client
      tags: [clients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Client"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          $ref: "#/components/responses/Problem"
  /v1/clients/{id}:
    get:
      operationId: getClient
      summary: Get client
      tags: [clients]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Client detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          $ref: "#/components/responses/Problem"
    patch:
      operationId: patchClient
      summary: Update client
      tags: [clients]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /v1/households:
    get:
      operationId: listHouseholds
      summary: List households
      tags: [households]
      responses:
        "200":
          description: Household list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
    post:
      operationId: createHousehold
      summary: Create household
      tags: [households]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Household"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
...
  /v1/profiles:
    get:
      operationId: listProfiles
      summary: List advisory profiles
      tags: [profiles]
      responses:
        "200":
          description: Profile list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
    post:
      operationId: createProfile
      summary: Create profile
      tags: [profiles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainRecord"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainRecord"
        "400":
          $ref: "#/components/responses/Problem"
...
  /v1/important-dates:
    get:
      operationId: listImportantDates
      summary: List important dates
      tags: [dates]
      responses:
        "200":
          description: Date list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
    post:
      operationId: createImportantDate
      summary: Create important date
      tags: [dates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainRecord"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainRecord"
        "400":
          $ref: "#/components/responses/Problem"
...
  /v1/addresses:
    get:
      operationId: listAddresses
      summary: List structured addresses
      tags: [addresses]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
    post:
      operationId: createAddress
      summary: Create address
      tags: [addresses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Address"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "400":
          $ref: "#/components/responses/Problem"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        total:
          type: integer
    Client:
      type: object
      required: [first_name, last_name, date_of_birth]
      properties:
        id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        address:
          type: string
        communication_preference:
          type: string
    ClientPatch:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        address:
          type: string
        communication_preference:
          type: string
    Household:
      type: object
      properties:
        id:
          type: string
        primary_client_id:
          type: string
        household_name:
          type: string
    DomainRecord:
      type: object
      additionalProperties: true
...
    Address:
      type: object
      required: [client_id, line1, suburb, state, postcode]
      properties:
        id:
          type: string
        client_id:
          type: string
        line1:
          type: string
        suburb:
          type: string
        state:
          type: string
        postcode:
          type: string
        country:
          type: string
    ProblemDetails:
      type: object
      required: [title, status, trace_id]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        trace_id:
          type: string
  responses:
    Problem:
      description: Problem details response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"